src/server/requirements.txt
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
flask-jwt-extended
flask-restful
flask-restful-swagger
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Injector==0.10.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
jsonref==0.2
celery==4.3.0
psycopg2==2.7.4

